Add files via upload
This zipped folder contains a python script for extracting occurrence data for avian species in the United States of America. Also in the folder is a README that explains how to use the python script and an "Avian Dictionary" for the python script to reference when searching for the species.